requirements.txt
streamlit
google-generativeai
pypdf
python-docx
streamlit-quill
beautifulsoup4
